// src/Query/enums/operator.ts
enum operatorEnum {
    EQUAL = '=',
    NOT_EQUAL = '<>',
    GREATER_THAN = '>',
    GREATER_THAN_OR_EQUAL = '>=',
    LESS_THAN = '<',
    LESS_THAN_OR_EQUAL = '<=',
    LIKE = 'like',
    NOT_LIKE = 'not like',
    IN = 'in',
    NOT_IN = 'not in',
    BETWEEN = 'between',
    NOT_BETWEEN = 'not between',
}


export default operatorEnum;